feat(navbar): add position and style for navbar
Set the background color using the CSS variable (--header-color) and
made the navbar fixed at the top of the page. Adjusted the font size
to 1.1rem and styled the nav ul to be a flex container, centering the
items. Added margin and a hover effect to the nav li elements. Ensured
that nav a elements have padding and inherit the text color.